Integration architecture: refactored

+++ app/src/main/java/com/mirogal/cocktail/presentation/receiver/ProposeDrinkReceiverLiveData.kt
@@ -6,21 +6,21 @@ import android.content.Intent
 import android.content.IntentFilter
 import androidx.lifecycle.LiveData
 
-class ProposeDrinkReceiverLiveData(val context: Context) : LiveData<Int?>() {
+class ProposeDrinkReceiverLiveData(val context: Context) : LiveData<Long?>() {
 
-    private var startId: Int = -1
+    private var startId: Long = -1
 
     private val receiver = object : BroadcastReceiver() {
         override fun onReceive(context: Context?, intent: Intent?) {
             val isTimerFinish: Boolean = intent?.getBooleanExtra("isTimerFinish", false) ?: false
             if (isTimerFinish) {
-                val finishId: Int = intent?.getIntExtra("finishCocktailId", -2) ?: -2
+                val finishId: Long = intent?.getLongExtra("finishCocktailId", -2L) ?: -2L
                 if (startId == finishId) {
                     value = finishId
                     value = null
                 }
             } else {
-                startId = intent?.getIntExtra("startCocktailId", -1) ?: -1
+                startId = intent?.getLongExtra("startCocktailId", -1L) ?: -1L
             }
         }
     }
